Add error reporting
Former-commit-id: 831f0d81248503a1a185c86c00034e519c11ed53

+++ plugins_install/diehard.sh
@@ -1,4 +1,10 @@
+#!/bin/bash
 
+if [ ! -f $CFAR_HOME/DieHard/src/libdiehard.so ]; then
+	echo "ERROR: DieHard library not built/found" | tee warning.txt
+	exit 1
+fi
 
-cp $CFAR_HOME/DieHard/src/libdiehard.so libheaprand.so
 $PEASOUP_HOME/tools/update_env_var.sh DO_DIEHARD 1
+cp $CFAR_HOME/DieHard/src/libdiehard.so libheaprand.so
+
